requirements.txt
python-whois==0.9.5
dnspython==2.7.0
requests==2.32.3
beautifulsoup4==4.12.3
shodan==1.31.0
jinja2==3.1.4
colorama==0.4.6
python-dotenv==1.0.1
tqdm==4.67.1
